requirements.txt
absl-py==0.12.0
aioredis==1.3.1
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
channels==3.0.1
channels-redis==2.4.2
chardet==4.0.0
click==8.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
Cython==0.29.21
daphne==3.0.1
Django==3.1.4
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
gensim==4.0.1
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hiredis==1.1.0
huggingface-hub==0.0.8
hyperlink==21.0.0
idna==2.10
image==1.5.33
importlib-metadata==4.0.1
incremental==17.5.0
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
msgpack==0.6.2
mysqlclient==2.0.3
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.1.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.5
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.22.2
scipy==1.6.3
seqeval==1.2.2
service-identity==18.1.0
six==1.15.0
sklearn==0.0
smart-open==5.0.0
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.6.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
underthesea==1.2.3
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==2.0.0
wrapt==1.12.1
zipp==3.4.1
zope.interface==5.2.0
